RESIST the ever-growing length of `infer/builder.rs`
